+++ features/tailormade_design_detail/src/main/java/com/future/tailormade_design_detail/core/model/response/SizeDetailResponse.kt
@@ -6,13 +6,13 @@ import kotlinx.android.parcel.Parcelize
 @Parcelize
 data class SizeDetailResponse(
 
-    val chest: Double? = null,
+    val chest: Float? = null,
 
-    val waist: Double? = null,
+    val waist: Float? = null,
 
-    val hips: Double? = null,
+    val hips: Float? = null,
 
-    val neckToWaist: Double? = null,
+    val neckToWaist: Float? = null,
 
-    val inseam: Double? = null,
+    val inseam: Float? = null,
 ): Parcelable
